refactor(settings): extract theme controls into dedicated components
Move dark mode and color theme controls into separate components.
Keep the settings dialog focused and easier to maintain.
