+++ src/components/settings/settings-dialog.tsx
@@ -1,14 +1,10 @@
-import { BgColorsOutlined, CheckOutlined, MoonOutlined, SunOutlined } from "@ant-design/icons";
-import { Divider, Modal, Space, Switch, Typography } from "antd";
+import { Divider, Modal, Typography } from "antd";
 import { createStyles } from "antd-style";
 
-import { useColorTheme, useIsDarkMode, useSetColorTheme, useThemeToggler } from "~/stores/settings.store";
-import { THEMES } from "~/themes";
-import type { ColorTheme } from "~/themes";
+import { SettingsColorTheme } from "./settings-color-theme";
+import { SettingsDarkLight } from "./settings-dark-light";
 
-const { Text, Title } = Typography;
-
-const THEME_KEYS = Object.keys(THEMES) as ColorTheme[];
+const { Title } = Typography;
 
 interface SettingsDialogProps {
   open: boolean;
@@ -17,10 +13,6 @@ interface SettingsDialogProps {
 
 export const SettingsDialog = ({ open, onClose }: SettingsDialogProps) => {
   const { styles } = useStyles();
-  const isDarkMode = useIsDarkMode();
-  const toggleThemeMode = useThemeToggler();
-  const colorTheme = useColorTheme();
-  const setColorTheme = useSetColorTheme();
 
   return (
     <Modal
@@ -35,54 +27,10 @@ export const SettingsDialog = ({ open, onClose }: SettingsDialogProps) => {
       width={400}
     >
       <Divider className={styles.divider} />
-
-      <Space orientation="horizontal" className={styles.row}>
-        <Space orientation="horizontal" align="center">
-          {isDarkMode ? <MoonOutlined className={styles.icon} /> : <SunOutlined className={styles.icon} />}
-          <Text strong>Dark Mode</Text>
-        </Space>
-
-        <Switch
-          checked={isDarkMode}
-          onChange={toggleThemeMode}
-          checkedChildren={<MoonOutlined />}
-          unCheckedChildren={<SunOutlined />}
-        />
-      </Space>
-
-      <Text type="secondary" className={styles.hint}>
-        Toggle between light and dark theme
-      </Text>
+      <SettingsDarkLight />
 
       <Divider className={styles.divider} />
-
-      <Space orientation="horizontal" className={styles.row}>
-        <Space orientation="horizontal" align="center">
-          <BgColorsOutlined className={styles.icon} />
-          <Text strong>Color Theme</Text>
-        </Space>
-
-        <Space size={8}>
-          {THEME_KEYS.map((key) => (
-            <button
-              key={key}
-              aria-label={THEMES[key].label}
-              className={styles.swatch}
-              style={{
-                backgroundColor: THEMES[key].primary,
-                boxShadow: colorTheme === key ? `0 0 0 2px ${THEMES[key].primary}` : undefined,
-              }}
-              onClick={() => setColorTheme(key)}
-            >
-              {colorTheme === key && <CheckOutlined className={styles.checkIcon} />}
-            </button>
-          ))}
-        </Space>
-      </Space>
-
-      <Text type="secondary" className={styles.hint}>
-        Choose an accent colour for the interface
-      </Text>
+      <SettingsColorTheme />
     </Modal>
   );
 };
@@ -93,34 +41,5 @@ const useStyles = createStyles(() => ({
   },
   divider: {
     margin: "12px 0 24px",
-  },
-  row: {
-    width: "100%",
-    justifyContent: "space-between",
-    alignItems: "center",
-  },
-  icon: {
-    fontSize: 20,
-  },
-  hint: {
-    display: "block",
-    marginTop: 8,
-    fontSize: 12,
-  },
-  swatch: {
-    width: 28,
-    height: 28,
-    borderRadius: "50%",
-    border: "2px solid transparent",
-    cursor: "pointer",
-    display: "inline-flex",
-    alignItems: "center",
-    justifyContent: "center",
-    padding: 0,
-    transition: "box-shadow 0.2s ease",
-  },
-  checkIcon: {
-    color: "#fff",
-    fontSize: 12,
   },
 }));
